+++ app/models/__init__.py
@@ -23,6 +23,7 @@ from .announcement import Announcement
 from .registration import Registration
 from .payment_event import PaymentEvent, record_payment_event
 from .rate_limit import RateBucket
+from .queue_stat import QueueStat
 from .issued_document import IssuedDocument
 from .abstract import Abstract, SPEAKER_STATUSES, SPEAKER_STATUS_ORDER, ALL_STATUSES
 from .organising_committee import OrganisingCommitteeMember
@@ -49,6 +50,7 @@ __all__ = [
     "Registration",
     "PaymentEvent", "record_payment_event",
     "RateBucket",
+    "QueueStat",
     "IssuedDocument",
     "Abstract", "SPEAKER_STATUSES", "SPEAKER_STATUS_ORDER", "ALL_STATUSES",
     "OrganisingCommitteeMember",